+++ CHANGES.rst
@@ -83,3 +83,4 @@ v1.0.7
 
 v1.0.8[unreleased]
 -----------
+* Allow members of organization to create/edit projects